requirements.txt
markdown
flask
flask_mySQL
